4032 - Move and update Admin/UserManagement to AdminCollaborators
