--- src/client/pages/Admin/UserManagement/index.ts
@@ -1,1 +1,0 @@
-export { default } from './UserManagement'
